Prevent 1/3-based line height values due to continued rounding errors
Line-heights of 1 + 1/3 are problematic in that the computer value will never
render to an exact pixel. The rounding error is workable at a font-size of 16,
but with a font-size of 15, Chrome computels the value to exactly 20 but then
rounds that value *down* to 19. This is odd because in other scenarios (like
with 16), the computed value contains enough decimal precision for the result
to be unnoticeable.

I looked into using big.js to get better decimal precision, but there was no
way to determine when to apply it and when not to. Rather than dealing with
that crazy math and an added dependency, I decided to encourage using 18 / 14
which is close enough to 1 + 1 / 3 for it not to really matter.