Refactor border compensation logic to use positioning offsets over margins
The issue with margins was when the first element on the page had a border,
the margin offset would push *down* the page instead of offsetting the
element. This was caused by margin collapsing. Specifically, the margin
couldn't be applied outside the boundaries of the body tag because the browser
doesn't allow that. Using position:relative allows the offset to move outside
of the constraints of the body tag and achieve the proper effect.