Remove all things related to colors
I decided color theory and managing color props fell outside the ideas driving
Stemcell so I axed them. This means that the color math is now gone along with
the ThemeProvider component as it is now un-necessary. It may make more sense
to push all of that logic out to a separate package.

The main motividation for this was that CSS color-mod via PostCSS doesn't work
with CSS variables. Adding support for them was non-trivial and ended up
making the logic in the Bit component to be a mess.

This commit *does* include a change from using Glamor with css() in the
className prop to using an array of object literals in a css prop so that they
can be converted into classNames at the core Bit component. Theoretically,
this allows usage of other css-in-js libraies other than Glamor.