Fix bugs in Rule that prevented VR alignment
Not only was marginRight being used to define marginLeft, the resulting
glamor classes had some specificity issues. The easiest way around it was to
manually define the margins in CSS as opposed to calculated values in Bit.

We have to manually calculate the vertical rhythm for the margin props
because of the math involved in calculating the line width. Size is
defined in pixels whereas marginBottom is defined in rhythm units. That
means that we pass in a calculated marginBottom offset into Bit. This
also means that we must manually calculate the other margin props because
as the left or right values change per instance from the user, it would
cause Glamor to generate a *new* class for the margin props but reuse a
previous rule for the color and offset. This causes the new box model
rules to have a higher CSS specificity which prevents the marginBottom
offset rule from overriding the base box props. Hopefully that makes
sense weeks from now...